Calibration write. tsl237 function returns int32 rather than float.

+++ Inc/tsl237.h
@@ -21,4 +21,5 @@ extern uint32_t tsl237t_done;
 void tsl237_vdd_on(void);
 void tsl237_vdd_off(void);
 void tsl237_command(char *);
-float tsl237_readsensor(void); 
+// float tsl237_readsensor(void);
+uint32_t tsl237_readsensor(void);
